EpochToSV to the drkey store prior to renaming
