+++ go/lib/drkey/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
         "//go/lib/infra/modules/db:go_default_library",
-        "//go/lib/log:go_default_library",
         "//go/lib/scrypto:go_default_library",
         "//go/lib/util:go_default_library",
         "@org_golang_x_crypto//pbkdf2:go_default_library",
